Save Button Disables when testing scene straight
