+++ P04-unity/P04 RTS SP/Assets/_Scripts/ChangeScenes.cs
@@ -1,0 +1,19 @@
+﻿using UnityEngine;
+using System.Collections;
+
+/// <summary>
+/// @Author: Andrew Seba
+/// @Description: Basic change to scene script.
+/// </summary>
+public class ChangeScenes : MonoBehaviour {
+
+	public void _MainScene()
+	{
+		Application.LoadLevel("MainGame");
+	}
+
+    public void _CreateGameScene()
+    {
+        Application.LoadLevel("CreateGameScene");
+    }
+}
