--- P04-unity/P04 RTS SP/Assets/_Scripts/ScriptInformationSave.cs
@@ -1,93 +1,0 @@
-﻿using UnityEngine;
-using System.IO;
-
-/// <summary>
-/// @Author: Andrew Seba
-/// @Description: takes the player and level name and writes it to a file under
-/// Assets/save_data/"levelName"/gameinformation.txt
-/// 
-/// Data Structure
-/// --------------
-/// userName
-/// levelName
-/// </summary>
-public class ScriptInformationSave : MonoBehaviour {
-
-    string userName = null;
-    string levelName = null;
-    string txtInfoLocation;
-    string outputFile = "/gameinformation.txt";
-
-    public string UserName
-    {
-        set
-        {
-            userName = value;
-        }
-    }
-
-    public string LevelName
-    {
-        set
-        {
-            levelName = value;
-        }
-    }
-
-    void Start()
-    {
-        DontDestroyOnLoad(gameObject);
-    }
-
-    /// <summary>
-    /// Takes the level name and makes a new folder.
-    /// Creates a gameinformation.txt file and writes the user name and level name.
-    /// </summary>
-	public void _InitFile()
-	{
-		txtInfoLocation = Application.dataPath + "/save_data/" + levelName;
-        StreamWriter writer = null;
-
-
-        //Check to see if both things have been entered as a safety.
-        if (userName != null && levelName != null)
-        {
-            if (!Directory.Exists(txtInfoLocation))
-            {
-                Directory.CreateDirectory(
-                    Application.dataPath + "/save_data/" + levelName);
-
-            }
-
-
-            //gameinformation.txt
-            //UserName
-            //LevelName
-            using (writer = new StreamWriter(txtInfoLocation + outputFile))
-            {
-                writer.WriteLine(userName);
-                writer.WriteLine(levelName);
-                
-            }
-            writer.Close();
-        }
-        else
-        {
-            Debug.Log("Either name or level name haven't been entered." +
-                "\nThis is required.");
-        }
-
-	}
-
-    public void SaveGame()
-    {
-        StreamWriter writer = null;
-        using (writer = new StreamWriter(txtInfoLocation + outputFile))
-        {
-            writer.WriteLine(userName);
-            writer.WriteLine(levelName);
-
-        }
-        writer.Close();
-    }
-}
