--- P04-unity/P04 RTS SP/Assets/_Scripts/ScriptStartGame.cs
@@ -1,10 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class ScriptStartGame : MonoBehaviour {
-
-	public void LoadMainScene()
-	{
-		Application.LoadLevel("MainGame");
-	}
-}
