+++ P04-unity/P04 RTS SP/Assets/_Scripts/SebaScripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9c35d9d8c3ad2d0408ac5381dc842a3f
+folderAsset: yes
+timeCreated: 1443887014
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
