+++ P04-unity/P04 RTS SP/Assets/_Scripts/SebaScripts/EnableDebugText.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b7f12bc4dfb4ecc46ac58f492eeae9f5
+timeCreated: 1443887024
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
