+++ P04-unity/P04 RTS SP/Assets/_Scripts/SebaScripts/GameInformation.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 28c61bc5eedae5649b545ab70d0d8d91
+timeCreated: 1443885896
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
