+++ P04-unity/P04 RTS SP/Assets/save_data/TestingSave.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7e82e9c3e45689843a5559394ab698b1
+folderAsset: yes
+timeCreated: 1443887773
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
