+++ P04-unity/P04 RTS SP/Assets/save_data/TestingSave/gameinformation.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 27216eaaff31eb549b54b79714c4e3fe
+timeCreated: 1443887773
+licenseType: Free
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
